requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/1db8ea6b-8e8d-11ee-9041-cedaeb4cabe2/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==22.0.0
h11==0.14.0
html5==0.0.9
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
packaging==23.2
pause==0.1.2
pycryptodome==3.20.0
PySocks==1.7.1
python-dateutil==2.7.3
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
selenium-stealth==1.0.6
six==1.11.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.18
watchdog==4.0.1
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.17.0
Flask==2.0.1
pycryptodome==3.10.1
gunicorn==20.1.0